Refactor: libpacemaker: move colocation functions to own source file
The colocation functions from pcmk_sched_constraints.c (which was >3K lines)
are now in a new file pcmk_sched_colocation.c.

There are no code changes aside from function renames, formatting, and
comments.
