+++ lib/pacemaker/pcmk_sched_group.c
@@ -15,6 +15,8 @@
 
 #include <pacemaker-internal.h>
 
+#include "libpacemaker_private.h"
+
 #define VARIANT_GROUP 1
 #include <lib/pengine/variant.h>
